Move some types to impl package
