+++ src/main/scala/singleton/ops/Concat.scala
@@ -2,7 +2,7 @@ package singleton.ops
 
 import macrocompat.bundle
 import scala.reflect.macros.whitebox
-import singleton.ops.macros.MacroUtils
+import singleton.ops.impl._
 
 trait Concat[A, B] extends Op
 
@@ -12,7 +12,7 @@ object Concat extends Op2Companion[Concat] {
     .materialize[A, B]
 
   @bundle
-  final class ConcatMacro(val c: whitebox.Context) extends MacroUtils {
+  final class ConcatMacro(val c: whitebox.Context) extends Macros {
     def materialize[A: c.WeakTypeTag, B: c.WeakTypeTag]: c.Tree =
       materializeBinaryOp[Concat, A, B].apply[String](_ + _)
   }
