+++ src/main/scala/singleton/ops/Times.scala
@@ -2,7 +2,7 @@ package singleton.ops
 
 import macrocompat.bundle
 import scala.reflect.macros.whitebox
-import singleton.ops.macros.MacroUtils
+import singleton.ops.impl._
 
 trait Times[A, B] extends Op
 
@@ -13,7 +13,7 @@ object Times extends Op2Companion[Times] {
   ): Times[A, B] = macro TimesMacro.materialize[T, A, B]
 
   @bundle
-  final class TimesMacro(val c: whitebox.Context) extends MacroUtils {
+  final class TimesMacro(val c: whitebox.Context) extends Macros {
     def materialize[T, A: c.WeakTypeTag, B: c.WeakTypeTag](
         nt: c.Expr[Numeric[T]]
     ): c.Tree =
